Changed to floats

+++ include/audio_processing.hpp
@@ -7,12 +7,12 @@
 #include "read_wav.hpp"
 
 std::vector<int16_t> convertBytes(const WavFile& wav);
-std::vector<std::vector<double>> sampleToBlock(std::vector<int16_t>& samples);
+std::vector<std::vector<float>> sampleToBlock(std::vector<int16_t>& samples);
 std::vector<std::complex<double>> discreteFourierTransform(
     const std::vector<double>& sample);
-std::vector<std::complex<double>> fastFourierTransform(
-    const std::vector<double>& sample);
+std::vector<std::complex<float>> fastFourierTransform(
+    const std::vector<float>& sample);
 std::vector<float> computeMagnitude(
-    const std::vector<std::complex<double>>& fourierResult);
+    const std::vector<std::complex<float>>& fourierResult);
 
 #endif
